agents modeling/mcp_server/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-generativeai
PyMuPDF